cilium: encryption, use fib_lookup to rewrite dmac/smac
Previously fib_lookup was only used if IP_POOLS was enabled. However,
we actually need to rewrite DMAC and SMAC whenever encrypt_node is
being used because in this mode we call redirect() from BPF side and
stack can not rewrite the fields.

This patch pulls the HAVE_FIB_LOOKUP block of BPF code outside the
IP_POOLS ifdef so that it is run in any case it is available.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>