test: add capability to specify images / skip K8s provisioning
Add three new options to the Ginkgo CI:

* `-cilium.operator-image` - image containing cilium-operator
* `-cilium.image` - image containing cilium-agent
* `-cilium.provision-k8s` - whether or not to provision K8s (true or false)

These flags are useful when running CI-only changes locally, so that developers
can quickly deploy cilium and cilium-operator without having to build it. This
also allows for tests to be ran against specific tags, e.g. v1.4.5, etc.

If either image option is not set, the scripts for running tests will build the
image which was not specified, and pull the other one. If neither are set, the
existing behavior of building both images is performed.

Signed-off by: Ian Vernon <ian@cilium.io>