eventqueue: block in `WaitToBeDrained` for in-flight events to be completed
Once an EventQueue has been closed, that means that no more events can be
Enqueued onto it. However, inflight events may still be running. Add a means of
signaling that an Event has been completed via a Semaphore, which
`WaitToBeDrained` tries to acquire after the queue has been closed. Once an
Event has finished, it will release its hold on the semaphore. There are
guaranteed to be no more Events which will be ran due to the queue being closed,
making acquisition of this semaphore a good signal that the event has completed.

Signed-off by: Ian Vernon <ian@cilium.io>